remove superfluous controller concern
